extract method to make testing easier
